somewhat works
